processing_service/requirements.txt
redis
python-dotenv
psycopg
psycopg-pool
requests
fastapi
pydantic